+++ scripts/create_conda_env.sh
@@ -85,7 +85,7 @@ esac
 
 echo "CONDA_OS=$CONDA_OS"
 
-CONDA_INSTALLER="Miniforge3-$CONDA_OS-$CONDA_ARCH.sh"
+CONDA_INSTALLER="Mambaforge-$CONDA_OS-$CONDA_ARCH.sh"
 
 echo "CONDA_INSTALLER=$CONDA_INSTALLER"
 
@@ -117,20 +117,20 @@ echo -e "\033[1;34mInstall compiler...\033[0m"
 
 case "$KERNEL" in
   Linux*)
-    "$CONDA_DIR/bin/conda" install -y "gxx_linux-$CONDA_PKG_EXT>=12"
+    "$CONDA_DIR/bin/mamba" install -y "gxx_linux-$CONDA_PKG_EXT>=12"
     ;;
   Darwin*)
-    "$CONDA_DIR/bin/conda" install -y "clang_osx-$CONDA_PKG_EXT>=16"
+    "$CONDA_DIR/bin/mamba" install -y "clang_osx-$CONDA_PKG_EXT>=16"
     ;;
 esac
 
 echo -e "\033[1;34mWORKAROUND\033[0m"
 
-"$CONDA_DIR/bin/conda" install -y 'conda==22.11.1'
+"$CONDA_DIR/bin/mamba" install -y 'conda==22.11.1'
 
 echo -e "\033[1;34mInstall toolchain...\033[0m"
 
-"$CONDA_DIR/bin/conda" install -y \
+"$CONDA_DIR/bin/mamba" install -y \
   clangdev \
   cmake \
   conda-build \
@@ -139,7 +139,7 @@ echo -e "\033[1;34mInstall toolchain...\033[0m"
 
 echo -e "\033[1;34mInstall dependencies...\033[0m"
 
-"$CONDA_DIR/bin/conda" install -y \
+"$CONDA_DIR/bin/mamba" install -y \
   benchmark \
   catch2 \
   jinja2 \
@@ -149,11 +149,11 @@ echo -e "\033[1;34mInstall dependencies...\033[0m"
 
 echo -e "\033[1;34mWORKAROUND\033[0m"
 
-"$CONDA_DIR/bin/conda" install -y 'conda==22.11.1'
+"$CONDA_DIR/bin/mamba" install -y 'conda==22.11.1'
 
 echo -e "\033[1;34mInstall dependencies from $BUILD...\033[0m"
 
-"$CONDA_DIR/bin/conda" install -y --channel "https://roq-trading.com/conda/$BUILD" \
+"$CONDA_DIR/bin/mamba" install -y --channel "https://roq-trading.com/conda/$BUILD" \
   roq-client \
   roq-tools